livekit_backend/pyproject.toml
[project]
name = "livekit-outbound-call"
version = "0.1.0"
description = "基于 LiveKit Agents 的智能外呼系统"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=0.12",
    "livekit-plugins-silero>=0.7",
    "livekit-plugins-openai>=0.10",
    "livekit-api>=0.7",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "hiredis>=2.0.0",
    "websockets>=12.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
